Added third subtypes

+++ app/Services/CharacterManager.php
@@ -111,6 +111,13 @@ class CharacterManager extends Service
                 if(!$subtypeTwo || $subtype->species_id != $data['species_id']) throw new \Exception('Selected secondary subtype invalid or does not match species.');
             }
             else $data['subtype_id_2'] = null;
+			if(isset($data['subtype_id_3']) && $data['subtype_id_3'])
+            {
+                $subtypeThree = Subtype::find($data['subtype_id_3']);
+                if(!(isset($data['species_id']) && $data['species_id'])) throw new \Exception('Species must be selected to select a subtype.');
+                if(!$subtypeThree || $subtype->species_id != $data['species_id']) throw new \Exception('Selected third subtype invalid or does not match species.');
+            }
+            else $data['subtype_id_3'] = null;
 
             // Get owner info
             $url = null;
@@ -190,6 +197,7 @@ class CharacterManager extends Service
                 $data['species_id'] = isset($data['species_id']) && $data['species_id'] ? $data['species_id'] : null;
                 $data['subtype_id'] = isset($data['subtype_id']) && $data['subtype_id'] ? $data['subtype_id'] : null;
                 $data['subtype_id_2'] = isset($data['subtype_id_2']) && $data['subtype_id_2'] ? $data['subtype_id_2'] : null;
+				$data['subtype_id_3'] = isset($data['subtype_id_3']) && $data['subtype_id_3'] ? $data['subtype_id_3'] : null;
                 $data['rarity_id'] = isset($data['rarity_id']) && $data['rarity_id'] ? $data['rarity_id'] : null;
             }
 
@@ -240,6 +248,7 @@ class CharacterManager extends Service
                 $data['species_id'] = (isset($data['species_id']) && $data['species_id']) ? $data['species_id'] : null;
                 $data['subtype_id'] = isset($data['subtype_id']) && $data['subtype_id'] ? $data['subtype_id'] : null;
                 $data['subtype_id_2'] = isset($data['subtype_id_2']) && $data['subtype_id_2'] ? $data['subtype_id_2'] : null;
+				$data['subtype_id_3'] = isset($data['subtype_id_3']) && $data['subtype_id_3'] ? $data['subtype_id_3'] : null;
                 $data['rarity_id'] = (isset($data['rarity_id']) && $data['rarity_id']) ? $data['rarity_id'] : null;
 
 
@@ -254,7 +263,7 @@ class CharacterManager extends Service
                 }
             }
             $imageData = Arr::only($data, [
-                'species_id', 'subtype_id', 'subtype_id_2', 'rarity_id', 'use_cropper',
+                'species_id', 'subtype_id', 'subtype_id_2', 'subtype_id_3', 'rarity_id', 'use_cropper',
                 'x0', 'x1', 'y0', 'y1',
             ]);
             $imageData['use_cropper'] = isset($data['use_cropper']) ;
@@ -640,6 +649,13 @@ class CharacterManager extends Service
                 if(!$subtypeTwo || $subtypeTwo->species_id != $data['species_id']) throw new \Exception('Selected secondary subtype invalid or does not match species.');
             }
             else $data['subtype_id_2'] = null;
+			if(isset($data['subtype_id_3']) && $data['subtype_id_3'])
+            {
+                $subtypeThree = Subtype::find($data['subtype_id_3']);
+                if(!(isset($data['species_id']) && $data['species_id'])) throw new \Exception('Species must be selected to select a subtype.');
+                if(!$subtypeThree || $subtypeThree->species_id != $data['species_id']) throw new \Exception('Selected third subtype invalid or does not match species.');
+            }
+            else $data['subtype_id_3'] = null;
 
             $data['is_visible'] = 1;
 
@@ -702,6 +718,12 @@ class CharacterManager extends Service
                 if(!(isset($data['species_id']) && $data['species_id'])) throw new \Exception('Species must be selected to select a subtype.');
                 if(!$subtypeTwo || $subtypeTwo->species_id != $data['species_id']) throw new \Exception('Selected secondary subtype invalid or does not match species.');
             }
+            if(isset($data['subtype_id_3']) && $data['subtype_id_3'])
+            {
+                $subtypeThree = Subtype::find($data['subtype_id_3']);
+                if(!(isset($data['species_id']) && $data['species_id'])) throw new \Exception('Species must be selected to select a subtype.');
+                if(!$subtypeThree || $subtypeThree->species_id != $data['species_id']) throw new \Exception('Selected third subtype invalid or does not match species.');
+            }
 
             // Log old features
             $old = [];
@@ -709,6 +731,7 @@ class CharacterManager extends Service
             $old['species'] = $image->species_id ? $image->species->displayName : null;
             $old['subtype'] = $image->subtype_id ? $image->subtype->displayName : null;
             $old['subtype_2'] = $image->subtype_id_2 ? $image->subtypeTwo->displayName : null;
+			$old['subtype_3'] = $image->subtype_id_3 ? $image->subtypeThree->displayName : null;
             $old['rarity'] = $image->rarity_id ? $image->rarity->displayName : null;
 
             // Clear old features
@@ -725,6 +748,7 @@ class CharacterManager extends Service
             $image->species_id = $data['species_id'];
             $image->subtype_id = $data['subtype_id'] ?: null;
             $image->subtype_id_2 = $data['subtype_id_2'] ?: null;
+			$image->subtype_id_3 = $data['subtype_id_3'] ?: null;
             $image->rarity_id = $data['rarity_id'];
             $image->save();
 
@@ -733,6 +757,7 @@ class CharacterManager extends Service
             $new['species'] = $image->species_id ? $image->species->displayName : null;
             $new['subtype'] = $image->subtype_id ? $image->subtype->displayName : null;
             $new['subtype_2'] = $image->subtype_id_2 ? $image->subtypeTwo->displayName : null;
+			$new['subtype_3'] = $image->subtype_id_3 ? $image->subtypeThree->displayName : null;
             $new['rarity'] = $image->rarity_id ? $image->rarity->displayName : null;
 
             // Character also keeps track of these features
@@ -1792,6 +1817,7 @@ is_object($sender) ? $sender->id : null,
                 'species_id' => $character->image->species_id,
                 'subtype_id' => $character->image->subtype_id,
                 'subtype_id_2' => $character->image->subtype_id_2,
+				'subtype_id_3' => $character->image->subtype_id_3,
             ];
 
             $request = CharacterDesignUpdate::create($data);
@@ -2059,10 +2085,14 @@ is_object($sender) ? $sender->id : null,
             if(isset($data['subtype_id_2']) && $data['subtype_id_2'])
                 $subtypeTwo = ($request->character->is_myo_slot && $request->character->image->subtype_id_2) ? $request->character->image->subtypeTwo : Subtype::find($data['subtype_id_2']);
             else $subtypeTwo = null;
+			if(isset($data['subtype_id_3']) && $data['subtype_id_3'])
+                $subtypeThree = ($request->character->is_myo_slot && $request->character->image->subtype_id_3) ? $request->character->image->subtypeThree : Subtype::find($data['subtype_id_3']);
+            else $subtypeThree = null;
             if(!$rarity) throw new \Exception("Invalid rarity selected.");
             if(!$species) throw new \Exception("Invalid species selected.");
             if($subtype && $subtype->species_id != $species->id) throw new \Exception("Subtype does not match the species.");
             if($subtypeTwo && $subtypeTwo->species_id != $species->id) throw new \Exception("Secondary subtype does not match the species.");
+			if($subtypeThree && $subtypeThree->species_id != $species->id) throw new \Exception("Third subtype does not match the species.");
 
             // Clear old features
             $request->features()->delete();
@@ -2090,6 +2120,7 @@ is_object($sender) ? $sender->id : null,
             $request->rarity_id = $rarity->id;
             $request->subtype_id = $subtype ? $subtype->id : null;
             $request->subtype_id_2 = $subtypeTwo ? $subtypeTwo->id : null;
+			$request->subtype_id_3 = $subtypeThree ? $subtypeThree->id : null;
             $request->has_features = 1;
             $request->save();
 
@@ -2211,6 +2242,7 @@ is_object($sender) ? $sender->id : null,
                 'species_id' => $request->species_id,
                 'subtype_id' => ($request->character->is_myo_slot && isset($request->character->image->subtype_id)) ? $request->character->image->subtype_id : $request->subtype_id,
                 'subtype_id_2' => ($request->character->is_myo_slot && isset($request->character->image->subtype_id_2)) ? $request->character->image->subtype_id_2 : $request->subtype_id_2,
+				'subtype_id_3' => ($request->character->is_myo_slot && isset($request->character->image->subtype_id_3)) ? $request->character->image->subtype_id_3 : $request->subtype_id_3,
                 'rarity_id' => $request->rarity_id,
                 'sort' => 0,
             ]);
